distro/rhel: move Azure dataloss warning script embed
Move the dataloss warning script contents to distro/rhel/ (out of rhel9)
so it can be used in rhel10 as well.
